+++ flockers/src/visualization/bird_vis.rs
@@ -22,8 +22,8 @@ impl AgentRender for BirdVis {
         let agent = agent.downcast_ref::<Bird>().unwrap();
         let loc = state.field1.get_location(*agent);
         match loc {
-            Some(loc) => (loc.x as f32, loc.y as f32, 0.),
-            None => (agent.loc.x as f32, agent.loc.y as f32, 0.),
+            Some(loc) => (loc.x, loc.y, 0.),
+            None => (agent.loc.x, agent.loc.y, 0.),
         }
     }
 
@@ -40,7 +40,7 @@ impl AgentRender for BirdVis {
         } else {
             concrete_agent.last_d.y.atan2(concrete_agent.last_d.x)
         };
-        (rotation + PI) as f32
+        rotation + PI
     }
 
     fn update(
